chore(pom.xml): add SCM configuration for git repository
The SCM configuration for the git repository was added to the pom.xml file. This allows for easier navigation to the repository and provides information about the connection and developer connection URLs. The URL was set to "https://github.com/liquibase/liquibase-percona.git" and the tag was set to "HEAD".